Do not raise an msg_fatal when weight sum is zero
If we perform an integration with multiple components, we might have
the situation that some of them are naturally zero, e.g. if a
soft mismatch component is evaluated, but the cms energy is not
enought to produce any resonances at all (in this case the real
component performs a usual FKS integration). Now, the same message
is displayed as a warning and the subsequent grid adaption is skipped.