Fix phs_fks_generator_2
The reference output has changed because of the typo fix
in d478ca459ee, which makes the program use the correct xi_max.
Supposedly, for the particular region used in the test, the ordering
of xi_plus and xi_minus has changed, leading to different values
of the final momenta.